Added collision with holes so that the game triggers a restart on marble loss.
This'll happen when no more marbles remain on the screen.
At the moment there is only 1 marble and the list of marbles can be found in MarbleManager. When adding a new marble we'll have to get that marble's gameobject into that list. (Could probably just look for any GameObject tagged with "Ball" on start TBH but that  can be done when we introduce multiple marbles)